Updated Gemfile to leverage Ruby version
Reduced maintenance burden for the team by picking up the Ruby version
from the `.ruby-version` file. 🎉 This change ensures the
`.ruby-version` file defines the canonical Ruby version for the project
as it *should* be the source of truth for the entire project.

Upgrading Ruby, in the future, will be easier by having one less file
to maintain. 😉