Added Tocer Rake support
Provides a convenient Rake task for auto-updating the README Table of
Contents. Here's how to find it:

    rake -T | grep toc

...and here's how to use it:

    rake toc

The development environment guard was put in place in order to not
accidentally attemp to load this Rake task in a non-development
environemnt so as not to cause issues with *test* or *production*
environments (as the gem wouldn't be loaded for those environments
either).